Solution
Basically since we cannot use a regular expression, I am comparing each character of the input string to each character of a string that contains all integers from 0-9. When there is one found we add the found integer to a List of characters and after iterating over the input string we join this output together into a string.